+++ packages/plugin-text/src/factory/editor.tsx
@@ -21,8 +21,6 @@ import { linkNode } from '../plugins/link'
 import { TextPluginOptions } from './types'
 import { isValueEmpty, TextPlugin, PluginRegistry } from '..'
 
-const emptyValue: Value = Value.fromJSON({})
-
 export const createTextEditor = (
   options: TextPluginOptions
 ): React.ComponentType<SlateEditorProps> => {
@@ -58,15 +56,11 @@ export const createTextEditor = (
           return { ...plugins[name], name }
         })
 
-    const [rawState, setRawState] = React.useState(emptyValue)
-
-    // run this effect once when mounted to load json value into slate
-    React.useEffect(() => {
+    const [rawState, setRawState] = React.useState(() => {
       // slate.js changed format with version 0.46
       // old format is still supported, but new states will be in new format
-      setRawState(Value.fromJSON(props.state.value))
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [])
+      return Value.fromJSON(props.state.value)
+    })
 
     const thisState = React.useRef(props.state)
     const lastValue = React.useRef(props.state.value)